Remove phpunit artefacts
